TB-1.2 se añade Seeder y se aplica la configuración web y Cross
Se implementa las configuraciones de la web y un primer usuario admin para manejar la aplicación

+++ api/src/main/java/com/jonaygr/api/Models/Registros/Ticket.java
@@ -16,7 +16,7 @@ public class Ticket extends RegistroBase{
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Id
     @Column(name = "ticket_id", nullable = false)
-    private int ticketId;
+    private Long ticketId;
 
     @ManyToMany(fetch = FetchType.EAGER)
     @JsonBackReference
